requirements.txt
torch
numpy
soundfile
moviepy
ffmpeg-python
gTTS
silero-vad
assemblyai
deep-translator
python-dotenv
